+++ modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/exec/rel/HashAggregateExecutionTest.java
@@ -21,11 +21,7 @@ import static org.apache.ignite.internal.sql.engine.exec.exp.agg.AggregateType.M
 import static org.apache.ignite.internal.sql.engine.exec.exp.agg.AggregateType.REDUCE;
 import static org.apache.ignite.internal.sql.engine.exec.exp.agg.AggregateType.SINGLE;
 import static org.apache.ignite.internal.util.CollectionUtils.first;
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
@@ -34,15 +30,10 @@ import org.apache.calcite.rel.RelCollation;
 import org.apache.calcite.rel.RelCollations;
 import org.apache.calcite.rel.core.AggregateCall;
 import org.apache.calcite.rel.type.RelDataType;
-import org.apache.calcite.sql.fun.SqlStdOperatorTable;
 import org.apache.calcite.util.ImmutableBitSet;
 import org.apache.calcite.util.ImmutableIntList;
 import org.apache.ignite.internal.sql.engine.exec.ExecutionContext;
 import org.apache.ignite.internal.sql.engine.exec.RowHandler;
-import org.apache.ignite.internal.sql.engine.type.IgniteTypeFactory;
-import org.apache.ignite.internal.sql.engine.util.TypeUtils;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
 
 /**
  * HashAggregateExecutionTest.
@@ -141,63 +132,5 @@ public class HashAggregateExecutionTest extends BaseAggregateTest {
         sort.register(aggRdc);
 
         return sort;
-    }
-
-
-    /**
-     * Test verifies that after rewind all groups are properly initialized.
-     */
-    @ParameterizedTest
-    @EnumSource
-    public void countOfEmptyWithRewind(TestAggregateType testAgg) {
-        ExecutionContext<Object[]> ctx = executionContext();
-        IgniteTypeFactory tf = ctx.getTypeFactory();
-        RelDataType rowType = TypeUtils.createRowType(tf, int.class, int.class);
-        ScanNode<Object[]> scan = new ScanNode<>(ctx, Collections.emptyList());
-
-        AggregateCall call = AggregateCall.create(
-                SqlStdOperatorTable.COUNT,
-                false,
-                false,
-                false,
-                ImmutableIntList.of(),
-                -1,
-                null,
-                RelCollations.EMPTY,
-                tf.createJavaType(int.class),
-                null
-        );
-
-        List<ImmutableBitSet> grpSets = List.of(ImmutableBitSet.of());
-
-        RelDataType aggRowType = TypeUtils.createRowType(tf, int.class);
-
-        SingleNode<Object[]> aggChain = createAggregateNodesChain(
-                testAgg,
-                ctx,
-                grpSets,
-                call,
-                rowType,
-                aggRowType,
-                rowFactory(),
-                scan
-        );
-
-        for (int i = 0; i < 2; i++) {
-            RootNode<Object[]> root = new RootNode<>(ctx) {
-                /** {@inheritDoc} */
-                @Override public void close() {
-                    // NO-OP
-                }
-            };
-
-            root.register(aggChain);
-
-            assertTrue(root.hasNext());
-            assertArrayEquals(row(0), root.next());
-            assertFalse(root.hasNext());
-
-            aggChain.rewind();
-        }
     }
 }
